+++ build.ps1
@@ -19,6 +19,9 @@ if ($env:appveyor){
 dotnet build -c Release Frank.sln /p:Version=$version$versionSuffix
 dotnet pack -c Release src/Frank /p:Version=$version$versionSuffix -o $psscriptroot/bin
 
+dotnet test test/Frank.Analyzers.Tests
+dotnet pack -c Release src/Frank.Analyzers /p:Version=$version$versionSuffix -o $psscriptroot/bin
+
 dotnet build -c Release Frank.Datastar.sln /p:Version=$version$versionSuffix
 dotnet test test/Frank.Datastar.Tests
 dotnet pack -c Release src/Frank.Datastar /p:Version=$version$versionSuffix -o $psscriptroot/bin